refactor(contracts): DXDVotingMachine from payable and signed genesis protocol implementations
