+++ test/helpers/constants.js
@@ -1,3 +1,3 @@
-const ARC_GAS_LIMIT = process.env.OVERRIDE_ARC_GAS_LIMIT || 6200000;
+const ARC_GAS_LIMIT = process.env.OVERRIDE_ARC_GAS_LIMIT || 9000000;
 
 module.exports = { ARC_GAS_LIMIT };
